requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
contourpy==1.3.3
cycler==0.12.1
dnspython==2.8.0
Flask==3.1.3
Flask-Login==0.6.3
Flask-PyMongo==3.0.1
fonttools==4.62.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
networkx==3.6.1
nltk==3.9.4
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pillow==12.2.0
pymongo==4.16.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
regex==2026.4.4
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
textblob==0.20.0
threadpoolctl==3.6.0
tqdm==4.67.3
urllib3==2.6.3
Werkzeug==3.1.8
